+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+* Add DraggableFlowGraphView to support draggable layout graph.
+* Abstract GraphView class.
+
+
 ## 0.0.1
 
 * Finish simple DAG layout